+++ CHANGES.md
@@ -7,3 +7,4 @@
 - Add documentation
 - Add FastAPI endpoint
 - Add software tests and CI configuration
+- Use CrateDB as job store